requirements.txt
requests
configparser
python-telegram-bot==22.5
urllib3==2.6.2
pymongo[srv]